Move DistanceMap to its own file
